refactor(popular): move under own folder
