+++ projects/client/src/lib/sections/lists/popular/PopularList.svelte
@@ -2,10 +2,9 @@
   import SectionList from "$lib/components/lists/section-list/SectionList.svelte";
   import type { MediaType } from "$lib/models/MediaType";
   import { UrlBuilder } from "$lib/utils/url/UrlBuilder";
-  import MediaItem from "./components/MediaItem.svelte";
-  import ViewAllButton from "./components/ViewAllButton.svelte";
-  import { usePopularList } from "./stores/usePopularList";
-  import { mediaListHeightResolver } from "./utils/mediaListHeightResolver";
+  import MediaItem from "../components/MediaItem.svelte";
+  import DrillableMediaList from "../drilldown/DrillableMediaList.svelte";
+  import { usePopularList } from "./usePopularList";
 
   type PopularListProps = {
     title: string;
