+++ projects/client/src/routes/movies/+page.svelte
@@ -5,7 +5,7 @@
   import * as m from "$lib/features/i18n/messages.ts";
   import RenderFor from "$lib/guards/RenderFor.svelte";
   import AnticipatedList from "$lib/sections/lists/anticipated/AnticipatedList.svelte";
-  import PopularList from "$lib/sections/lists/PopularList.svelte";
+  import PopularList from "$lib/sections/lists/popular/PopularList.svelte";
   import RecommendedList from "$lib/sections/lists/RecommendedList.svelte";
   import TrendingList from "$lib/sections/lists/TrendingList.svelte";
   import {
