+++ projects/client/src/routes/movies/popular/+page.svelte
@@ -1,7 +1,7 @@
 <script>
   import TraktPage from "$lib/components/layout/TraktPage.svelte";
   import * as m from "$lib/features/i18n/messages";
-  import PopularPaginatedList from "$lib/sections/lists/PopularPaginatedList.svelte";
+  import PopularPaginatedList from "$lib/sections/lists/popular/PopularPaginatedList.svelte";
 
   import { DEFAULT_SHARE_MOVIE_COVER } from "$lib/utils/constants";
 </script>
