+++ projects/client/src/routes/shows/popular/+page.svelte
@@ -2,7 +2,7 @@
   import TraktPage from "$lib/components/layout/TraktPage.svelte";
   import * as m from "$lib/features/i18n/messages";
 
-  import PopularPaginatedList from "$lib/sections/lists/PopularPaginatedList.svelte";
+  import PopularPaginatedList from "$lib/sections/lists/popular/PopularPaginatedList.svelte";
   import { DEFAULT_SHARE_SHOW_COVER } from "$lib/utils/constants";
 </script>
